+++ test/generators/install_generator_test.rb
@@ -36,7 +36,7 @@ class InstallGeneratorTest < Rails::Generators::TestCase
     if RUBY_ENGINE == "jruby"
       assert_file "config/initializers/sequel.rb", /Sequel\.connect\("jdbc:sqlite:\/\/", extensions: :activerecord_connection\)/
     else
-      assert_file "config/initializers/sequel.rb", /Sequel\.sqlite\(extensions: :activerecord_connection\)/
+      assert_file "config/initializers/sequel.rb", /Sequel\.connect\("sqlite:\/\/", extensions: :activerecord_connection\)/
     end
   end